Hibernate migrated to version 5.6.15 (was 4.3.11)  - using Infinispan as second-level cache    - has JGroups-based clustering out of the box, but needs a newver version of JGroups  - using C3P0 database connection pool  - various changes across the whole codebase
- Other changes
  - MySQL: using TableGenerator with legacy-hilo optimization algoritghm (instead of the TableHiLoGenerator)
  - Location.permanentId changed to be Long (instead of Integer)
  - some code is deprecated
    - org.hibernate.Query -> org.hibernate.query.Query (with different setters)
    - orh.hibernate.Criteria is deprecated (and does not exist in Hibernate 6)
  - ClassMetadata is deprecated (still used), and a MappingException is thrown when the class is not present
  - setFlushMode -> setHibernateFlushMode
  - SessionFactory.getAllClassMetadata() no longer works
  - HQL: Location.pictures or PreferenceGroup.classes no longer works, but Location.buildingAbbv or Location.name are still fine

+++ JavaSource/org/unitime/timetable/model/base/BaseInstructionalOffering.java
@@ -40,7 +40,7 @@ public abstract class BaseInstructionalOffering implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	private Long iUniqueId;
-	private Integer iInstrOfferingPermId;
+	private Long iInstrOfferingPermId;
 	private Boolean iNotOffered;
 	private Integer iDemand;
 	private Integer iEnrollment;
@@ -91,8 +91,8 @@ public abstract class BaseInstructionalOffering implements Serializable {
 	public Long getUniqueId() { return iUniqueId; }
 	public void setUniqueId(Long uniqueId) { iUniqueId = uniqueId; }
 
-	public Integer getInstrOfferingPermId() { return iInstrOfferingPermId; }
-	public void setInstrOfferingPermId(Integer instrOfferingPermId) { iInstrOfferingPermId = instrOfferingPermId; }
+	public Long getInstrOfferingPermId() { return iInstrOfferingPermId; }
+	public void setInstrOfferingPermId(Long instrOfferingPermId) { iInstrOfferingPermId = instrOfferingPermId; }
 
 	public Boolean isNotOffered() { return iNotOffered; }
 	public Boolean getNotOffered() { return iNotOffered; }
